feat(nextjs): migrate to next
