+++ .gitignore
@@ -2,3 +2,5 @@ node_modules
 dist
 /.parcel-cache
 .env
+.env.local
+/.next
