+++ package.json
@@ -3,30 +3,22 @@
   "version": "1.0.0",
   "description": "",
   "scripts": {
-    "format": "prettier --write \"src/**/*.js\"",
-    "start": "parcel index.html",
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "parcel build index.html"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "buffer": "^6.0.3",
-    "crypto-browserify": "^3.12.0",
-    "events": "^3.3.0",
-    "os-browserify": "^0.3.0",
-    "parcel": "^2.9.3",
-    "path-browserify": "^1.0.1",
-    "postcss": "^8.4.24",
-    "prettier": "^3.0.3",
-    "process": "^0.11.10",
-    "stream-browserify": "^3.0.0",
-    "tailwindcss": "^3.3.2"
+    "dev": "next dev",
+    "start": "next start",
+    "build": "next build"
   },
   "dependencies": {
-    "dotenv": "^16.1.4",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "@types/node": "20.8.0",
+    "@types/react": "18.2.23",
+    "@types/react-dom": "18.2.8",
+    "autoprefixer": "10.4.16",
+    "eslint": "8.50.0",
+    "eslint-config-next": "13.5.3",
+    "next": "13.5.3",
+    "postcss": "8.4.31",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "tailwindcss": "3.3.3",
+    "typescript": "5.2.2"
   }
 }
