--- src/utils/links.js
@@ -1,74 +1,0 @@
-export const links = [
-  {
-    url: "https://images.unsplash.com/photo-1475070929565-c985b496cb9f?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=2070&q=80",
-    title: "Nature",
-    description:
-      "<p>Through photography, the beauty of Mother Nature can be frozen in time This category celebrates the magic of our planet and beyond - from the vastness of the outdoors to miraculous moments in your own backyard<p>",
-  },
-  {
-    url: "https://images.unsplash.com/photo-1465479423260-c4afc24172c6?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=2069&q=80",
-    title: "Shoes",
-    description:
-      "<p>Shoes Photography captures the essence of style and footwear, where each step becomes a unique statement. This category celebrates the world of shoes, from fashionable trends to the artistry of craftsmanship, all waiting to be explored one step at a time.<p>",
-  },
-  {
-    url: "https://images.unsplash.com/photo-1502982720700-bfff97f2ecac?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=2070&q=80",
-    title: "Camera",
-    description:
-      "<p>Explore the world through the lens of a camera. This category embraces the art of photography, from capturing breathtaking landscapes to the fine details of everyday life, all frozen in time for your viewing pleasure.<p>",
-  },
-  {
-    url: "https://images.unsplash.com/photo-1599676609657-b4235ea50e6d?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=2070&q=80",
-    title: "Birds",
-    description:
-      "<p>Witness the fascinating world of our feathered friends. This category celebrates the beauty of avian creatures, their graceful flights, vibrant plumage, and the unique moments they create in the wild.<p>",
-  },
-  {
-    url: "https://images.unsplash.com/photo-1512149673953-1e251807ec7c?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=2070&q=80",
-    title: "Phones",
-    description:
-      "<p>Discover the ever-evolving technology in the palm of your hand. This category showcases the world of smartphones, from cutting-edge innovations to their role in modern communication and connectivity.<p>",
-  },
-  {
-    url: "https://images.unsplash.com/photo-1481018085669-2bc6e4f00eed?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=2070&q=80",
-    title: "Movies",
-    description:
-      "<p>Let's pay homage to the beginnings of photography by celebrating the analogue of the past and present. From vintage polaroids to mesmerizing 35mm photos - this category examines the best of what film photography has to offer.<p>",
-  },
-  {
-    url: "https://images.unsplash.com/photo-1510906594845-bc082582c8cc?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=2044&q=80",
-    title: "Science",
-    description:
-      "<p>Dive into the realms of discovery and innovation. This category explores the wonders of science, from groundbreaking research to the captivating mysteries of the natural world and beyond.<p>",
-  },
-  {
-    url: "https://images.unsplash.com/photo-1479142506502-19b3a3b7ff33?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=2070&q=80",
-    title: "History",
-    description:
-      "<p>Journey through the annals of time. This category delves into the rich tapestry of human history, showcasing pivotal moments, iconic figures, and the fascinating stories that have shaped our world.<p>",
-  },
-  {
-    url: "https://images.unsplash.com/photo-1499856871958-5b9627545d1a?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=2020&q=80",
-    title: "Travel",
-    description:
-      "<p>Travel allows us to explore the world's wonders and experience new cultures. It's a journey of discovery, from iconic landmarks to hidden gems. Pack your bags and set out on an adventure to create memories that will last a lifetime.<p>",
-  },
-  {
-    url: "https://images.unsplash.com/photo-1493514789931-586cb221d7a7?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=2071&q=80",
-    title: "Busines",
-    description:
-      "<p>In the world of business, innovation and entrepreneurship drive progress. Whether you're a startup founder or a seasoned executive, the business landscape offers challenges and opportunities. Success comes to those who adapt, innovate, and lead.<p>",
-  },
-  {
-    url: "https://images.unsplash.com/photo-1491841550275-ad7854e35ca6?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=1974&q=80",
-    title: "People",
-    description:
-      "<p>People are the heart of our world, with diverse stories, experiences, and perspectives. From everyday heroes to global leaders, individuals shape our societies. Together, we strive for equality, understanding, and a brighter future.<p>",
-  },
-  {
-    url: "https://images.unsplash.com/photo-1528357136257-0c25517acfea?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=2070&q=80",
-    title: "Religion",
-    description:
-      "<p>Religion provides spiritual guidance and a sense of purpose for billions around the globe. It's a source of strength, faith, and community. Through rituals, teachings, and beliefs, people find meaning in their lives and a connection to the divine.<p>",
-  },
-];
