--- tailwind.config.js
@@ -1,14 +1,0 @@
-/** @type {import('tailwindcss').Config} */
-module.exports = {
-  content: [
-    "./src/**/*.{html,js}",
-  ],
-  theme: {
-    extend: {
-      height: {
-        '128': '37rem',
-      }
-    },
-  },
-  plugins: [],
-}
